order-service/requirements.txt
Flask==2.3.3
flask-cors==4.0.1
google-cloud-pubsub